feat(orchestrator): enforce node timeouts and whole-incident watchdog
Enforce architecture §2.9 execution limits with per-node timeouts and an async whole-incident watchdog. Nodes exceeding their budget raise NodeTimeoutError and route to failure handling. If the incident budget elapses, the watchdog cancels in-flight work, performs emergency twin teardown and PagerDuty escalation, and marks the incident ESCALATED.

Implements build-plan step 5.6.

+++ src/understudy/common/config.py
@@ -34,11 +34,11 @@ class TimeoutSettings(BaseModel):
 
     model_config = ConfigDict(frozen=True)
 
-    fork_seconds: int = 120
-    candidate_apply_seconds: int = 60
-    observation_seconds: int = 200
-    kernel_seconds: int = 5
-    incident_seconds: int = 600
+    fork_seconds: float = 120.0
+    candidate_apply_seconds: float = 60.0
+    observation_seconds: float = 200.0
+    kernel_seconds: float = 5.0
+    incident_seconds: float = 600.0
 
 
 class ClusterSettings(BaseModel):
